--- a/scipy/spatial/distance.py
+++ b/scipy/spatial/distance.py
@@ -105,6 +105,7 @@ __all__ = [
 ]
 
 
+import math
 import warnings
 import numpy as np
 import dataclasses
@@ -637,9 +638,9 @@ def correlation(u, v, w=None, centered=True):
     uv = np.dot(u, vw)
     uu = np.dot(u, uw)
     vv = np.dot(v, vw)
-    dist = 1.0 - uv / np.sqrt(uu * vv)
+    dist = 1.0 - uv / math.sqrt(uu * vv)
     # Return absolute value to avoid small negative value due to rounding
-    return np.abs(dist)
+    return abs(dist)
 
 
 def cosine(u, v, w=None):