@lgeiger 2023-11-24
From what I can tell the CI failure seems to be unrelated but happy to have a closer look if you think that's not the case.