@lucascolley 2023-11-24
> From what I can tell the CI failure seems to be unrelated but happy to have a closer look if you think that's not the case.

Yeah, that's unrelated.

Out of interest, is `u @ w` any faster/slower than `np.dot(u, w)`?