@tylerjereddy 2023-11-24
and thanks for your first contribution to SciPy